Moved decorators added node change command
