+++ .vscode/settings.json
@@ -10,6 +10,14 @@
     "[Dockerfile]": {
         "editor.formatOnSave": false
     },
-    "editor.tabSize": 4
-
+    "editor.tabSize": 4,
+    "files.exclude": {
+        "**/.git": true,
+        "**/.svn": true,
+        "**/.hg": true,
+        "**/CVS": true,
+        "**/.DS_Store": true,
+        "**/*.pyc": {"when": "$(basename).py"}, 
+        "**/__pycache__": true,
+    },
 }
