+++ cogs/music/decorators.py
